Update 2._Scope.md
The scope of this Working Group is to specify the format of the Lottie Animation Format, both the allowed JSON structure, the meaning of that structure, and how such Lottie files are to be rendered. The definition of the correct rendering will be specified by a combination of verbiage in the written specification and by exemplar Lottie files and their desired renderings committed to the `tests` repo.